+++ CHANGELOG.md
@@ -7,10 +7,19 @@
 - Baywatch default parameters added: `CB_BAYWATCH_ENABLED` and`CB_BAYWATCH_EXTERN_LOCATION`
 - Logs are saved via lospout
 - TLS client certificate needed by Cloudbreak is generated with `cbd generate`
+- Command `aws delete-role ` added
+- Command `aws generate-role` added
+- Command `aws show-role` added
+- Command `cloudbreak-shell` added
+- Command `cloudbreak-shell-quiet` added
+- Command `local-dev` added
+- Command `token` added
 
 ### Removed
 
 ### Changed
+
+- AWS authentication env varibale is fixed to use the correct AWS_SECRET_ACCESS_KEY (instead the old AWS_SECRET_KEY)
 
 ## [v0.5.3] - 2015-06-03